매장 목록 수정

+++ src/test/java/com/zerobase/reserve/domain/common/constants/ReviewConstants.java
@@ -3,5 +3,5 @@ package com.zerobase.reserve.domain.common.constants;
 public class ReviewConstants {
     public static final Long ID = 1L;
     public static final String CONTENTS = "맛있어요!";
-    public static final Integer SCORE = 5;
+    public static final Double SCORE = 5.0;
 }
